requirements.txt
python-jose[cryptography]
passlib[bcrypt]
fastapi
uvicorn
pydantic
python-multipart